Expose sub components out for a more composition based approach to using the lib.
