+++ src/components/SeatMap/api.js
@@ -1,5 +1,4 @@
-import { JetsApiService } from '../../common/api';
-import { DEFAULT_LANG, DEFAULT_UNITS } from '../../common';
+import { DEFAULT_LANG, DEFAULT_UNITS, JetsApiService } from '../../common';
 
 const API_SUPPORTED_LANGUAGES = [
   'AR',
