+++ src/components/index.js
@@ -1,0 +1,16 @@
+export * from './Bulk';
+export * from './Button';
+export * from './Deck';
+export * from './DeckExit';
+export * from './DeckSelector';
+export * from './Demo';
+export * from './NoData';
+export * from './Nose';
+export * from './NotInit';
+export * from './PlaneBody';
+export * from './Row';
+export * from './Seat';
+export * from './SeatMap';
+export * from './Tail';
+export * from './TooltipGlobal';
+export * from './Wing';
